+++ consumers/testing/src/testing/__init__.py
@@ -1,3 +1,3 @@
-"""__init__ for pytack consumer testing package."""
+"""__init__ for pkgdx consumer testing package."""
 
 __all__: list[str] = []